+++ cpp/dolfinx/geometry/CMakeLists.txt
@@ -8,5 +8,4 @@ set(HEADERS_geometry
 
 target_sources(dolfinx PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/gjk.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/utils.cpp
 )
